requirements.txt
pyautogui>=0.9.53
pygetwindow>=0.0.9
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0

